pyproject.toml
[project]
name = "pbl-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pygame-ce>=2.5.6",
    "pytmx>=3.32",
]
